feat: use double for measurements, set default serial /dev/USB0 @11250, update interp/filter/serial/util

+++ src/filter.h
@@ -1,5 +1,7 @@
 #ifndef FILTER_H
 #define FILTER_H
-// simple moving average filter
-void moving_average_3(const int *in, int n, double *out);
+
+// moving average over 3 Werte (in/out double)
+void moving_average_3(const double *in, int n, double *out);
+
 #endif
